+++ lib/src/generator/generate_translations.dart
@@ -112,9 +112,10 @@ void _generateClass(
     // even if this attribute exist, it has to satisfy the same signature as
     // specified in the interface
     // this error seems to occur when using in combination with "extends"
-    final optional = node.interface?.attributes.any((attribute) =>
-                attribute.optional && attribute.attributeName == key) ==
-            true
+    final optional = config.fallbackStrategy == FallbackStrategy.baseLocale &&
+            node.interface?.attributes.any((attribute) =>
+                    attribute.optional && attribute.attributeName == key) ==
+                true
         ? '?'
         : '';